+++ EasyBay/EasyBay/Interfaces/IBusinessLogic.cs
@@ -8,12 +8,12 @@ namespace EasyBay.Interfaces
 {
     public interface IBusinessLogic
     {
-        void RaisePrice(User user, Lot lot, int newPrice);
+        void RaisePrice(User user, Lot lot, decimal newPrice);
 
         void BuyOut(User user, Lot lot);
 
         void ProcessPurchase(Lot lot);
 
-        void Deposit(User user, int amount);
+        void Deposit(User user, decimal amount);
     }
 }
